LPS-186252 Rename and move to internal.cache directory
